pre/web/server/requirements.txt
absl-py==2.3.1
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.17.1
aniso8601==9.0.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
bidict==0.23.1
blinker==1.7.0
cachetools==6.2.1
certifi==2023.11.17
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
dnspython==2.4.2
email-validator==2.1.0.post1
eventlet==0.40.3
filelock==3.13.1
Flask==2.3.3
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Migrate==4.1.0
flask-restx==1.2.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==25.9.23
fonttools==4.46.0
frozenlist==1.4.0
fsspec==2023.12.1
gast==0.4.0
google-auth==2.41.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.0.1
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
idna==3.6
importlib-resources==6.1.1
itsdangerous==2.1.2
jax==0.4.28
jaxlib==0.4.28
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-specifications==2023.11.2
keras==2.12.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
libclang==18.1.1
Mako==1.3.10
Markdown==3.9
MarkupSafe==2.1.3
matplotlib==3.8.2
mediapipe==0.10.9
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
numpy==1.23.5
# nvidia-cublas-cu12==12.1.3.1
# nvidia-cuda-cupti-cu12==12.1.105
# nvidia-cuda-nvrtc-cu12==12.1.105
# nvidia-cuda-runtime-cu12==12.1.105
# nvidia-cudnn-cu12==8.9.2.26
# nvidia-cufft-cu12==11.0.2.54
# nvidia-curand-cu12==10.3.2.106
# nvidia-cusolver-cu12==11.4.5.107
# nvidia-cusparse-cu12==12.1.0.106
# nvidia-nccl-cu12==2.18.1
# nvidia-nvjitlink-cu12==12.9.86
# nvidia-nvtx-cu12==12.1.105
oauthlib==3.3.1
opencv-contrib-python==4.11.0.86
opencv-python==4.8.1.78
opt_einsum==3.4.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
protobuf==3.20.3
psutil==5.9.6
psycopg2-binary==2.9.9
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.8.0
pyparsing==3.1.1
pyrsistent==0.20.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.12.3
python-socketio==5.14.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.31.1
reportlab==4.0.8
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.13.2
rsa==4.9.1
scipy==1.15.3
seaborn==0.13.0
sentencepiece==0.2.1
simple-websocket==1.1.0
six==1.16.0
sounddevice==0.5.3
SQLAlchemy==2.0.23
sympy==1.12
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
thop==0.1.1.post2209072238
tomli==2.3.0
torch==2.1.1
torchvision==0.16.1
tqdm==4.66.1
triton==2.1.0
twilio==8.10.3
typing_extensions==4.15.0
tzdata==2023.3
ultralytics==8.0.222
urllib3==2.1.0
Werkzeug==2.3.8
wrapt==1.14.1
wsproto==1.2.0
yarl==1.9.3
